lens: move integration tests to separate file
